# ARROW-8199: [C++] Add support for multi-column sort indices on Table