--- a/c_glib/arrow-glib/version.h.in
+++ b/c_glib/arrow-glib/version.h.in
@@ -111,6 +111,15 @@
 #endif
 
 /**
+ * GARROW_VERSION_3_0:
+ *
+ * You can use this macro value for compile time API version check.
+ *
+ * Since: 3.0.0
+ */
+#define GARROW_VERSION_3_0 G_ENCODE_VERSION(3, 0)
+
+/**
  * GARROW_VERSION_2_0:
  *
  * You can use this macro value for compile time API version check.
@@ -228,6 +237,20 @@
 
 
 #define GARROW_AVAILABLE_IN_ALL
+
+#if GARROW_VERSION_MIN_REQUIRED >= GARROW_VERSION_3_0
+#  define GARROW_DEPRECATED_IN_3_0                GARROW_DEPRECATED
+#  define GARROW_DEPRECATED_IN_3_0_FOR(function)  GARROW_DEPRECATED_FOR(function)
+#else
+#  define GARROW_DEPRECATED_IN_3_0
+#  define GARROW_DEPRECATED_IN_3_0_FOR(function)
+#endif
+
+#if GARROW_VERSION_MAX_ALLOWED < GARROW_VERSION_3_0
+#  define GARROW_AVAILABLE_IN_3_0 GARROW_UNAVAILABLE(3, 0)
+#else
+#  define GARROW_AVAILABLE_IN_3_0
+#endif
 
 #if GARROW_VERSION_MIN_REQUIRED >= GARROW_VERSION_2_0
 #  define GARROW_DEPRECATED_IN_2_0                GARROW_DEPRECATED