--- a/cpp/src/arrow/compute/api_vector.cc
+++ b/cpp/src/arrow/compute/api_vector.cc
@@ -54,10 +54,18 @@ Result<std::shared_ptr<Array>> SortIndices(const Array& values, SortOrder order,
   return result.make_array();
 }
 
-Result<std::shared_ptr<Array>> SortIndices(const Table& values,
-                                           const SortOptions& options, ExecContext* ctx) {
+Result<std::shared_ptr<Array>> SortIndices(const ChunkedArray& chunked_array,
+                                           SortOrder order, ExecContext* ctx) {
+  SortOptions options({SortKey("not-used", order)});
+  ARROW_ASSIGN_OR_RAISE(
+      Datum result, CallFunction("sort_indices", {Datum(chunked_array)}, &options, ctx));
+  return result.make_array();
+}
+
+Result<std::shared_ptr<Array>> SortIndices(const Table& table, const SortOptions& options,
+                                           ExecContext* ctx) {
   ARROW_ASSIGN_OR_RAISE(Datum result,
-                        CallFunction("sort_indices", {Datum(values)}, &options, ctx));
+                        CallFunction("sort_indices", {Datum(table)}, &options, ctx));
   return result.make_array();
 }