… c_glib/doc/arrow-glib/arrow-glib-docs.xml
… c_glib/test/test-sort-indices.rb
… ci/conda_env_cpp.yml
… cpp/src/arrow/compute/kernels/vector_sort_benchmark.cc
… cpp/src/arrow/compute/kernels/vector_sort_test.cc
… cpp/thirdparty/versions.txt
… docs/source/cpp/compute.rst