@cyb70289 2020-11-11
Also update kernel document? `docs/source/cpp/compute.rst`