@pitrou 2020-11-24
Thanks for the numbers. I didn't expect "first sort key" to be better, but I'm convinced now.